Split oversized spawn test file into focused modules
